api/requirements.txt
django
djangorestframework
django-cors-headers
requests
celery
redis
python-dotenv
gunicorn
